feat: Add scale factor to serve metrics calculation
The code changes in this commit add a scale factor to the calculation of serve metrics. The scale factor is retrieved from the configuration file and defaults to 1.0 if not specified. This allows for more flexibility in adjusting the metrics based on the scale of the video.

This commit also updates the logging message to reflect the changes made in the focused serve analysis report.

```