feat: Refactor serve_analysis/utils.py and serve_analysis/constants.py
The code changes in this commit refactor the `calculate_angle` function in `serve_analysis/utils.py` to accept a list of floats instead of a list of lists. This simplifies the code and improves performance.

In addition, the `TENNIS_KEYPOINTS` dictionary in `serve_analysis/constants.py` has been updated to include additional keypoint mappings.

These changes improve the efficiency and readability of the code in the serve analysis module.